ran-guardian/pyproject.toml
[tool.poetry]
name = "ran-guardian"
version = "0.1.0"
description = ""
authors = ["Amit Verma <amiacs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-genai = "^0.5.0"
google-cloud-datastore = "^2.20.2"
googlemaps = "^4.10.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.5"
streamlit = "^1.41.1"
streamlit-calendar = "^1.2.1"
google-cloud-firestore = "^2.20.0"
pandas-gbq = "^0.26.1"
google-cloud-storage = "^2.19.0"
sse-starlette = "^2.2.1"
folium = "^0.19.4"
streamlit-folium = "^0.24.0"
google-cloud-bigquery-storage = "^2.27.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
lab = "^8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
